requirements.txt
langchain-openai==0.3.1
openai==1.60.0
streamlit==1.41.1
python-dotenv==1.0.1
langgraph==0.2.67
graphviz==0.20.3
python-chess